refactor: share handle drawing and pixbuf upload helpers
Move the crop tool's handle drawing into drag_box and turn the
renderer-private background upload into a reusable
create_image_from_pixbuf, so other tools can draw handles and upload
pixbufs to the canvas.
